refactor: split referer patch by platform
