+++ .github/workflows/test.yml
@@ -39,12 +39,14 @@ jobs:
         with:
           distribution: 'zulu'
           java-version: '21'
+      - name: Smoke test Capacitor Referer patch on Android
+        run: bun run test:patch:smoke:android
       - name: Build
         id: build_code
-        run: npm run verify:android
+        run: bun run verify:android
       - name: Run Android Tests
         id: test_android
-        run: npm run test:android
+        run: bun run test:android
   build_ios:
     timeout-minutes: 30
     runs-on: macOS-latest
@@ -57,12 +59,14 @@ jobs:
         run: bun i
       - name: Check plugin wiring
         run: bun run check:wiring
+      - name: Smoke test Capacitor Referer patch on iOS
+        run: bun run test:patch:smoke:ios
       - name: Build
         id: build_code
         run: bun run verify:ios
       - name: Run iOS Tests
         id: test_ios
-        run: npm run test:ios
+        run: bun run test:ios
   web:
     timeout-minutes: 30
     runs-on: ubuntu-latest
@@ -76,7 +80,7 @@ jobs:
         run: bun i
       - name: Check plugin wiring
         run: bun run check:wiring
-      - name: Test Capacitor Referer patcher
+      - name: Test Capacitor Referer patcher fixtures
         run: bun run test:patch
       - name: Lint
         id: lint_code
