+++ package.json
@@ -36,24 +36,26 @@
   "scripts": {
     "capacitor:sync:before": "node ./scripts/apply-capacitor-youtube-referer-patch.mjs",
     "capacitor:update:before": "node ./scripts/apply-capacitor-youtube-referer-patch.mjs",
-    "verify": "npm run verify:ios && npm run verify:android && npm run verify:web",
+    "verify": "bun run verify:ios && bun run verify:android && bun run verify:web",
     "verify:ios": "xcodebuild -scheme CapgoCapacitorYoutubePlayer -destination generic/platform=iOS",
     "verify:android": "cd android && ./gradlew clean build test && cd ..",
-    "verify:web": "npm run build",
-    "test": "npm run test:ios && npm run test:android",
+    "verify:web": "bun run build",
+    "test": "bun run test:ios && bun run test:android",
     "test:patch": "node --test scripts/test-capacitor-referer-patch.mjs",
+    "test:patch:smoke:android": "node scripts/test-capacitor-referer-smoke.mjs android",
+    "test:patch:smoke:ios": "node scripts/test-capacitor-referer-smoke.mjs ios",
     "test:ios": "./scripts/test-ios.sh",
     "test:android": "cd android && ./gradlew test && cd ..",
-    "lint": "npm run eslint && npm run prettier -- --check && npm run swiftlint -- lint",
-    "fmt": "npm run eslint -- --fix && npm run prettier -- --write && npm run swiftlint -- --fix --format",
+    "lint": "bun run eslint && bun run prettier -- --check && bun run swiftlint -- lint",
+    "fmt": "bun run eslint -- --fix && bun run prettier -- --write && bun run swiftlint -- --fix --format",
     "eslint": "eslint . --ext ts",
     "prettier": "prettier-pretty-check \"**/*.{css,html,ts,js,java}\" --plugin=prettier-plugin-java",
     "swiftlint": "node-swiftlint",
     "docgen": "docgen --api YoutubePlayerPlugin --output-readme README.md --output-json dist/docs.json",
-    "build": "npm run clean && npm run docgen && tsc && rollup -c rollup.config.mjs",
+    "build": "bun run clean && bun run docgen && tsc && rollup -c rollup.config.mjs",
     "clean": "rimraf ./dist",
     "watch": "tsc --watch",
-    "prepublishOnly": "npm run build",
+    "prepublishOnly": "bun run build",
     "check:wiring": "node scripts/check-capacitor-plugin-wiring.mjs"
   },
   "devDependencies": {
